requirements.txt
aiogram
fastapi
pydantic
python-dotenv
pillow
sqlite-utils
uvicorn
sqlite3
httpx